moved editable-cell-fields.jsx to grid-edit folder for better management
